renamed templates client to EmailTemplatesAPI, updated yard annotations, added http clients memoization
