+++ lib/mailtrap.rb
@@ -4,6 +4,6 @@ require_relative 'mailtrap/action_mailer' if defined? ActionMailer
 require_relative 'mailtrap/mail'
 require_relative 'mailtrap/errors'
 require_relative 'mailtrap/version'
-require_relative 'mailtrap/template'
+require_relative 'mailtrap/email_template'
 
 module Mailtrap; end
